Fixing warnings. Tried faster install, but this ate up too much memory. See http://www.boost.org/doc/libs/1_42_0/libs/python/doc/tutorial/doc/html/python/techniques.html#python.reducing_compiling_time

+++ extensions/crystal_ext.cpp
@@ -117,7 +117,7 @@ bp::list _GetScatteringComponentList(Crystal &c)
 {
     const ScatteringComponentList& scl = c.GetScatteringComponentList();
     bp::list l;
-    for(size_t i = 0; i < scl.GetNbComponent(); ++i)
+    for(int i = 0; i < scl.GetNbComponent(); ++i)
     {
         l.append(scl(i));
     }
